changed correct cypress file name
